requirements.txt
isort>5.1.0,<6.0.0
# black~=24.1

nemo_toolkit[all]==1.21.0
torch
omegaconf
torchmetrics
pytorch_lightning
jiwer
